requirements.txt
Pillow
mysql-connector-python
beautifulsoup4